requirements.txt
amqp==5.0.6
billiard==3.6.4.0
boto3==1.18.56
botocore==1.21.56
celery==5.1.2
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.7.0
django-storages==1.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.0.4
idna==2.10
kombu==5.1.0
libmagic==1.0
pilkit==2.0
Pillow==8.1.0
psycopg2-binary==2.9.1
python-magic==0.4.18
pytz==2020.4
redis==3.5.3
s3transfer==0.5.0
vine==5.0.0
whitenoise==5.2.0

